backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
skyfield==1.46
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.26.2
asyncpg==0.29.0
aioredis==2.0.1
fastapi-cors==0.0.6
websockets==12.0


